pyproject.toml
[project]
name = "adk-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.2.1",
    "litellm>=1.72.1",
    "openai>=1.84.0",
    "python-dotenv>=1.1.0",
    "yfinance>=0.2.61",
]
